feat: handle private buckets (#43)
* chore: add new bucket to example site

* feat: use AWS pre-signed URL to get objects from private buckets

* test: add e2e test to check that the example site has 1502 images

* chore: remove region plugin option

The option is not required anymore when using pre-signed URLs

* revert: "chore: remove region plugin option"

This reverts commit d870d3c44be22a2505326a7454dd46dd91410d85.

The AWS SDK apparently still expects a region.

* chore(scripts): add local start script with additional env cleanup

* docs: add READMEs to the starter and testing directories

* docs(tests): add details to the testing README

* fix: fix testing folder structure

Cypress sees any file in the integrationFolder as a spec. The spec had to be nexted one level deeper to allow for the README

* docs: update README

* feat: get signed URL only for image files in onCreateNode

BREAKING CHANGE: the Url key doesn't exist on s3Object anymore. There is not a public URL for private buckets and file nodes are now sourced with pre-signed URLs that shouldn't be added to the schema.