chore: change env var names to ensure config comes from plugin options
If the AWS credentials are not the env vars that AWS expects, requests fail (bad request). In some cases, the environment variables can not be the ones that AWS expects, because they are reserved (by Netlify, for example).

This means that the config doesn't come from the plugin, as it should, and falls back on the environment variables. Changing the names in the example ensures that this will not happen in the future.