fix: use plugin options for AWS config (#45)
* chore: change env var names to ensure config comes from plugin options

If the AWS credentials are not the env vars that AWS expects, requests fail (bad request). In some cases, the environment variables can not be the ones that AWS expects, because they are reserved (by Netlify, for example).

This means that the config doesn't come from the plugin, as it should, and falls back on the environment variables. Changing the names in the example ensures that this will not happen in the future.

* fix: revert to getting pre-signed URL in sourceNodes

The AWS config wasn't properly passed in onCreateNode, and the plugin therefore relied on the AWS official environment variables being set (as a fallback). This commit fixes it, but it also reintroduces the unwanted url key on s3Object. This will be addressed in a follow-up PR.

* fix: pass new env var names to GitHub Actions workflow